# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloe API Documentation",
        "description": "Cloe API documentation",
        "contact": {
            "name": "Jesus Romero",
            "email": "gisuss8520@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://cloeBE.test",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/utils/contact": {
            "post": {
                "tags": [
                    "Utils"
                ],
                "summary": "Contacto con centro de soporte de cloe",
                "description": "Contacto con centro de soporte de cloe.",
                "operationId": "baafb2fef727f402af03af4eb136a72e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email enviado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/utils/estados": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Estados registrados en sistema cloe",
                "description": "Estados registrados en sistema cloe.",
                "operationId": "b065ebe1e88880bd40bd1d5680c67a7c",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/utils/ciudades": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Ciudades registrados en sistema cloe",
                "description": "Ciudades registrados en sistema cloe.",
                "operationId": "8794eb36c35280ea78acab9d97cb35c2",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/utils/municipios": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Municipios registrados en sistema cloe",
                "description": "Municipios registrados en sistema cloe.",
                "operationId": "ad6b8d054eb38ed2717c13aa4c7f383d",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/cargos/": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Cargos registrados en sistema cloe",
                "description": "Cargos registrados en sistema cloe.",
                "operationId": "b10cebc970b28aa7a59f3a0f8b1b56be",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/utils/brands": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Marcas de RAEES pre-cargados en sistema cloe",
                "description": "Marcas de RAEES pre-cargados en sistema cloe.",
                "operationId": "5c1fdcc1566eaaecfdb65527e00ab53e",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/utils/lines": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Líneas de clasificación de RAEES.",
                "description": "Líneas de clasificación de RAEES.",
                "operationId": "863dbf8b02038176d34bdbcf4dfc0d84",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/utils/categories": {
            "get": {
                "tags": [
                    "Utils"
                ],
                "summary": "Categorías de RAEES.",
                "description": "Categorías de RAEES.",
                "operationId": "8f697c3d6855d60ddc34e94f9b0afbb1",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ContactRequest": {
                "properties": {
                    "name": {
                        "description": "Nombre",
                        "type": "string",
                        "example": "John Doe",
                        "nullable": false
                    },
                    "phone": {
                        "description": "Telefono",
                        "type": "string",
                        "example": "04121234567",
                        "nullable": false
                    },
                    "email": {
                        "description": "Correo electronico",
                        "type": "string",
                        "example": "johndoe@example.com",
                        "nullable": false
                    },
                    "city": {
                        "description": "Ciudad",
                        "type": "string",
                        "example": "Valencia",
                        "nullable": false
                    },
                    "message": {
                        "description": "Mensaje",
                        "type": "string",
                        "example": "Centro de acopio en la zona de valencia?",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}